Move resizing functionality to dedicated trait and class
So that we can avoid using libraries using java.awt.*, that make
native-image crash on Windows for now.
